获取股票详情

+++ models/tag.go
@@ -63,8 +63,8 @@ func GetTags(pageNum int, pageSize int, maps interface{}) ([]Tag, error) {
 }
 
 // GetTagTotal counts the total number of tags based on the constraint
-func GetTagTotal(maps interface{}) (int, error) {
-	var count int
+func GetTagTotal(maps interface{}) (int64, error) {
+	var count int64
 	if err := db.Model(&Tag{}).Where(maps).Count(&count).Error; err != nil {
 		return 0, err
 	}
